Create flexible driver for find_approximate_matches
The format of this driver should be pretty good as a general-purpose driver
right out of the box for most remaining problems. Certainly, it will need a
few tweaks right where it calls the targeted routine, but that's to be
expected. The boilerplate of handling the HTML example, the sample dataset or
the actual dataset has been separated out. If there's sample output, it's
compared to the actual output, and an error message is created if they don't
match. If there's no sample output, the result is written to the screen for
pasting into the Stepic text box.